graphics: fix ScrollRaster/ScrollRasterBF dropping negative deltas (aarch64)
graphics.conf declares both functions with WORD parameters, so every
caller stub passes 16-bit arguments, but the implementations declared
them AROS_LHA(LONG, ...). On aarch64 the libcall macros are plain C
calls under AAPCS64: the caller loads a WORD from memory with ldrh
(zero-extend) and is not required to sign-extend narrow arguments, so
the callee reading a LONG sees dy = -30 arrive as 65506 (0xFFE2).

The huge |dy| then always takes the "scrolled outside the rect" bail:
EraseRect + return, no blit. Under Zune windows the layer backfill is
LAYERS_NOBACKFILL, so the erase is a no-op too and a negative scroll
does nothing except the caller's exposed-strip redraw. Every wheel-up
or leftwards scroll left the rest of the view un-moved, which showed
up as duplicated icon rows / stale bands in Wanderer drawer windows
(and compounded: later positive scrolls faithfully blit the mess).

Fix: make the implementation parameter types match the .conf (WORD),
so the callee sign-extends the low 16 bits itself. This also keeps
every already-built caller binary correct.

Positive deltas always worked, which is why the corruption looked
resize-related: resizes are what commonly clamp the view offset and
produce negative scroll deltas.

Note: a sweep found the same latent WORD(.conf)/LONG(impl) mismatch in
BltBitMap, RectFill, EraseRect, Flood, Read/WritePixelLine8,
WriteChunkyPixels and WritePixelArray8 coordinates; those are almost
always non-negative in practice and are left for a follow-up.

Diagnosed by logging the delta at the IconList call site vs inside
ScrollRasterBF: d=(0,-30) left the app, d=(0,65506) arrived.

+++ rom/graphics/scrollrasterbf.c
@@ -18,12 +18,12 @@
 
 /*  SYNOPSIS */
         AROS_LHA(struct RastPort *, rp, A1),
-        AROS_LHA(LONG             , dx, D0),
-        AROS_LHA(LONG             , dy, D1),
-        AROS_LHA(LONG             , xMin, D2),
-        AROS_LHA(LONG             , yMin, D3),
-        AROS_LHA(LONG             , xMax, D4),
-        AROS_LHA(LONG             , yMax, D5),
+        AROS_LHA(WORD             , dx, D0),
+        AROS_LHA(WORD             , dy, D1),
+        AROS_LHA(WORD             , xMin, D2),
+        AROS_LHA(WORD             , yMin, D3),
+        AROS_LHA(WORD             , xMax, D4),
+        AROS_LHA(WORD             , yMax, D5),
 
 /*  LOCATION */
         struct GfxBase *, GfxBase, 167, Graphics)
